backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2026.1.1
mozilla-django-oidc==4.0.1
django-cors-headers==4.4.0
django-environ==0.11.2
dj-database-url==2.2.0
psycopg[binary]==3.2.1
cryptography==43.0.3
whitenoise==6.7.0
gunicorn==22.0.0
Markdown==3.7
bleach==6.1.0
